requirements.txt
openai
pandas
duckdb
pydantic
python-dotenv